second commit

+++ guru/index.php
@@ -53,7 +53,7 @@ if (!isset($_SESSION['status']) && !isset($_SESSION['username'])) {
                 <a class="btn btn-warning btn-sm" href='edit.php?nip=<?= $row_guru['nip'] ?>'>
                   Edit
                 </a>
-                <a class="btn btn-danger btn-sm" href='delete.php?nip=<?= $row_guru['nip'] ?>'>
+                <a class="btn btn-danger btn-sm" href='delete.php?nip=<?= $row_guru['nip'] ?>' onclick="return confirm('Anda yakin ingin menghapus data ini?');">
                   Delete
                 </a>
               </td>
